requirements.txt
streamlit
Pillow
ollama
requests